хостинг-maxhost/push-server/requirements.txt
pywebpush>=1.14.0
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
